remove dropping empty rows from read_csv_to_dataframe to get_validated_dataframe
